# Fuse q k v gemm to a kernel