This PR merge Q and KV into a single linear layer to reduce the number of GEMM operations, and thus reduce the kernel launch overheads.

Please have a review  @zhuohan123 @WoosukKwon