@zhaoyang-star 2023-08-30
This PR will cause GPU block drop. But thanks to the decrease in latency, the throughput will increase slightly.

Test env:
- A100 40 GB
- Model: starcoder

| Items | GPU Blocks | Online Throughput(request_rate=1, in_len/out_len=1k, tp=1) |
|---|---|---|
| Baseline | 18766 | 1241.78 tokens/sec 
| + PR | 18638 | 1246.69 tokens/sec

| Items | GPU Blocks | Latency(bs=1, in_len/out_len=1k, num_iters=10, tp=1) |
|---|---|---|
| Baseline | 20946 | 30.2455 sec | 
| + PR |  20818 | 29.8238 sec |