@zhuohan123 2023-11-21
This is fixed in #1622. Thank you for your contribution!